+++ tests/testthat/test-corners.R
@@ -4,7 +4,7 @@ test_that("empty anonymouse functions", {
 
   # This breaks when you mess with the order of the extract_docstrings function
   expect_equal(16 %>>% (function(x){}) %>% esc, NULL)
-  expect_equal(as_monad({}) %>% esc, NULL)
+  expect_equal(evalwrap({}) %>% esc, NULL)
 
   # This just can never work, because bind wants to pass a value
   # It might be preferable to catch this and throw a special error
@@ -20,5 +20,5 @@ test_that("NULL can be value", {
 })
 
 test_that("Error messages can be empty", {
-  expect_true(as_monad(stop()) %>% has_error)            
+  expect_true(evalwrap(stop()) %>% has_error)            
 })