+++ tests/testthat/test-monad-laws.R
@@ -7,7 +7,7 @@ context("monad laws")
 
 # into `(a -> b)`, which is provided by the user, and the internal evaluation
 # function `(a -> b) -> a -> m b`, which evaluates the functional application,
-# wrapping the result in an Rmonad object. This second function is `as_monad`.
+# wrapping the result in an Rmonad object. This second function is `evalwrap`.
 
 test_that("First law", {
   # left identity
@@ -15,16 +15,16 @@ test_that("First law", {
   # ------------------------
 
   # rmonad is a bit sloppy in this case. If the LFS is not an Rmonad, it first
-  # evaluates it into an Rmonad with `as_monad`.
-  # So the expression, `16 %>>% sqrt` is identical to `as_monad(16) %>>% sqrt`.
-  expect_true(rmonad_equal(16 %>>% sqrt, as_monad(16) %>>% sqrt))
+  # evaluates it into an Rmonad with `evalwrap`.
+  # So the expression, `16 %>>% sqrt` is identical to `evalwrap(16) %>>% sqrt`.
+  expect_true(rmonad_equal(16 %>>% sqrt, evalwrap(16) %>>% sqrt))
 
   # The way bind works in rmonad is a bit odd, the Haskell bind signature is:
   #   bind :: m a -> (a -> m b) -> m b
   # But the rmonad signature is, superficially, closer to:
   #   bindR :: m a -> (a -> b) -> m b
 
-  # However, internally `as_monad` takes the lhs `a` and the rhs `(a -> b)` and
+  # However, internally `evalwrap` takes the lhs `a` and the rhs `(a -> b)` and
   # applies them to get `m b`. Effectively: `(a -> (a -> b)) -> m b`. Thus,
   # overall, bindR has the proper signature.
 
@@ -41,7 +41,7 @@ test_that("First law", {
 #   # m >>= return  ===  m
 #
 #   # NOTE: Unfortunately, this currently isn't quite true
-#   # expect_true(rmonad_equal(16 %>>% sqrt, 16 %>>% sqrt %>>% as_monad))
+#   # expect_true(rmonad_equal(16 %>>% sqrt, 16 %>>% sqrt %>>% evalwrap))
 # })
 #
 # test_that("Third law", {